Refactor digitiser_data
