+++ trace-writer/src/error.rs
@@ -13,7 +13,5 @@ pub(crate) enum TraceWriterError {
     TimestampConversionFailed,
 
     #[error("Timestamp count not be converted to nanoseconds since epoch: {0}")]
-    NanosecondConversionFailed(DateTime<Utc>),
-    //#[error("String could not be written as HDF5 unicode: {0}")]
-    //UnicodeConversionFailed(String),
+    NanosecondConversionFailed(DateTime<Utc>)
 }
